refactor: ScanRepeater -> Repeater로 변경. 컨트롤 로직에 활용
- 컨트롤 로직에서도 충분히 활용 가능한 메소드를 담고 있었기 때문에 이를 컨트롤러에도 적용했습니다.
- 동시에 View뿐만 아니라 Controller에도 적용하는 유틸리티의 성격이 강하므로 패키지를 변경하였습니다.
