+++ src/main/java/util/Repeater.java
@@ -1,16 +1,18 @@
-package view;
+package util;
+
+import view.OutputView;
 
 import java.util.function.Supplier;
 
-public class ScanRepeater {
+public class Repeater {
     private static OutputView outputView = OutputView.getInstance();
 
-    public static <T> T scanOrRepeatOnError(Supplier<T> supplier) {
+    public static <T> T repeatFunctionOnError(Supplier<T> supplier) {
         try {
             return supplier.get();
         } catch (RuntimeException e) {
             outputView.printError(e);
-            return scanOrRepeatOnError(supplier);
+            return repeatFunctionOnError(supplier);
         }
     }
 }
